feat(report): 生产任务进度添加 mqtt 实时更新
- 新增 mqtt 服务配置和连接逻辑- 实现了生产任务进度的实时更新功能
- 优化了数据刷新和展示逻辑
- 添加了扫码枪类型判断和相应数据更新

+++ lib/bean/http/response/sap_wms_reprint_label_info.dart
@@ -1,6 +1,5 @@
-
 class ReprintLabelInfo {
-  bool select=false;
+  bool select = false;
 
   String? factory; //工厂  WERKS
   String? warehouse; //仓库  LGORT
@@ -14,7 +13,7 @@ class ReprintLabelInfo {
   String? instructionNo; //指令号 ZVBELN_ORI
   String? salesOrderNumber; //销售订单号  KDAUF
   int? salesOrderLineItem; //销售订单行号  KDPOS
-  int? quantity; //数量 MENGE
+  double? quantity; //数量 MENGE
   String? unit; // MEINS
   String? deliveryDate; //单位 EINDT
   String? numPage; //交货日期  ZPQYM
@@ -65,7 +64,7 @@ class ReprintLabelInfo {
     instructionNo = json['ZVBELN_ORI'];
     salesOrderNumber = json['KDAUF'];
     salesOrderLineItem = json['KDPOS'];
-    quantity = json['MENGE'];
+    quantity = json['MENGE'].toDouble();
     unit = json['MEINS'];
     deliveryDate = json['EINDT'];
     numPage = json['ZPQYM'];
